Add files via upload
Predictive Analytics on Global Terrorism Dataset from Kaggle using Python and Machine Learning

The world is experiencing an era where terrorist attacks are becoming prevalent in various parts of the globe. Every nation is trying hard to strengthen their counter terrorism strategies, policies and defense operations to tackle the menace of terrorism. The world is reeling from a wave of deadly terrorist attacks. The prime responsibility of every nation is to provide security to its citizens from such attacks. Any attack adversely affects the growth of a nation to a large extent which requires a lot of resources to compensate the loss.

The following are the main objectives of the project

To perform exploratory analysis to find patterns and trends in the dataset which would be useful for further analyzing process.

To predict the terrorist organization behind the attack using parameters such as the weapons used, the area of attack and the type of target chosen for the attack.

To predict the kind of weapons used by analyzing the organization involved along with the attack region and casualties.

To evaluate different classifiers and compare the results of different models.